refactor : divide components into smaller components
