+++ apps/web/app/examples/members/components/features/index.ts
@@ -1,0 +1,4 @@
+export { default as OurCulture } from './our-culture';
+export { default as RecentActivity } from './recent-activity';
+export { default as TeamGrowth } from './team-growth';
+export { default as TeamMembers } from './team-members';
