+++ apps/web/app/examples/members/components/index.ts
@@ -1,3 +1,2 @@
 export { default as ManageMember } from './manage-member';
 export { default as HireMember } from './hire-member';
-export { default as RecentActivity } from './recent-activity';
