+++ apps/web/app/examples/members/page.tsx
@@ -1,15 +1,12 @@
-import {
-  ManageMember,
-  HireMember,
-  RecentActivity,
-} from '@/app/examples/members/components';
+import { ManageMember, HireMember } from '@/app/examples/members/components';
+import { OurCulture } from '@/app/examples/members/components/features';
 
 export default function LandingPage() {
   return (
     <div className='flex flex-nowrap gap-4 p-5'>
       <div className='flex w-4/12 flex-col gap-y-6'>
         <ManageMember />
-        <RecentActivity />
+        <OurCulture />
       </div>
       <HireMember />
     </div>
